replace Text with TextTtf, added TextOtf
