+++ src/meson.build
@@ -76,7 +76,8 @@ source_file = [
     'StrokeLine.cpp',
     'StrokeMiterlimit.cpp',
     'Svg.cpp',
-    'Text.cpp',
+    'TextOtf.cpp',
+    'TextTtf.cpp',
     'TextEffects.cpp',
     'TextLayout.cpp',
     'TextLineWrap.cpp',
